added new styles for frontend

+++ src/main/java/org/knowledgeroot/app/page/PageDao.java
@@ -16,7 +16,7 @@ public interface PageDao {
      * @param id
      * @return
      */
-    Page findById(long id);
+    Page findById(Integer id);
 
     /**
      * check if page exists
